fix(bindings): un-skip spanset.test + geoset.test with signature fixes
Two more wholesale-skipped files cleared.

* **spanset.test** — the shared issue across `width(intspanset, bool)`,
  `duration(datespanset, bool)`, `duration(tstzspanset, bool)` was that
  `SpansetFunctions::Numspanset_width` / `Datespanset_duration` /
  `Tstzspanset_duration` each read the BOOLEAN second argument with
  `FlatVector::GetData<int32_t>(...)`, tripping 1.4's type-match assert
  ("Expected INT32, found BOOL"). Read as `bool` to match the declared
  `LogicalType::BOOLEAN`.

* **geoset.test** — `SpatialSetFunctions::Set_num_values` wrote the
  result as `int64_t` although `numValues` is registered as returning
  `LogicalType::INTEGER` (INT32). The sibling
  `SetFunctions::Set_num_values` already had it right; align the
  spatial version.

Remove the wholesale `mode skip` wrapper from each file. Only one
bump-PR skip remains: `test/sql/tbox.test` (several tbox constructor
signatures still need reviewing).

+++ src/geo/geoset.cpp
@@ -393,7 +393,9 @@ void SpatialSetFunctions::Set_num_values(DataChunk &args, ExpressionState &state
     auto &input = args.data[0];
     input.Flatten(args.size());
     auto input_data = FlatVector::GetData<string_t>(input);
-    auto result_data = FlatVector::GetData<int64_t>(result);
+    // numValues is registered as returning INTEGER (INT32); DuckDB 1.4 asserts
+    // the result Vector's template type matches the declared type.
+    auto result_data = FlatVector::GetData<int32_t>(result);
 
     for (idx_t i = 0; i < args.size(); i++) {
         if (FlatVector::IsNull(input, i)) {
